requirements.txt
langchain==0.3.23
langchain-core==0.3.51
langchain-community==0.3.21 
langchain-text-splitters==0.3.8 
pydantic-settings==2.8.1
langcodes==3.5.0
langgraph==0.3.30
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langid==1.1.6
langsmith==0.3.8
language_data==1.3.0
google==2.0.3
google-api-core==1.34.1
google-api-python-client==2.160.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.79.0
google-cloud-automl==1.0.1
google-cloud-bigquery==3.25.0
google-cloud-bigquery-connection==1.17.0
google-genai==1.7.0
google-ai-generativelanguage==0.6.17 
langchain-google-genai==2.1.2
pydantic==2.11.3
pydantic_core==2.33.1
pydata-google-auth==1.9.1
httpcore==1.0.7
httpimport==1.4.0
httplib2==0.22.0
httpx==0.28.1
python-dotenv
#google-generativeai - DO NOT INSTALL
